+++ warden/app/app.go
@@ -877,6 +877,12 @@ func NewApp(
 	}
 
 	app.UpgradeKeeper.SetUpgradeHandler("v0.7.1", func(ctx context.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+		sdkCtx := sdk.UnwrapSDKContext(ctx)
+
+		if err := app.EVMKeeper.AddPreinstalls(sdkCtx, evmtypes.DefaultPreinstalls); err != nil { //nolint
+			return nil, fmt.Errorf("enable preinstalls: %w", err)
+		}
+
 		return app.ModuleManager.RunMigrations(ctx, app.Configurator(), fromVM)
 	})